pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = "A simple Diary application for the paranoid"
authors = [{ name = "Lorenzo", email = "lorenzofumi@gmail.com" }]
license = { text = "Apache 2.0" }
readme = "README.md"
requires-python = ">=3.12.3,<3.15"
dependencies = [
  "pyqt6 (>=6.9.1,<7.0.0)",
  "pydantic-settings (>=2.11.0,<3.0.0)",
  "pytest (>=8.4.2,<9.0.0)",
  "argon2-cffi (>=25.1.0,<26.0.0)",
  "pynacl (>=1.6.0,<2.0.0)",
  "msgpack (>=1.1.2,<2.0.0)",
  "zstd (>=1.5.7.2,<2.0.0.0)",
  "paddleocr (>=2.7.0,<3.0.0)",
  "pytesseract (>=0.3.10,<0.4.0)",
  "pillow (>=10.0.0,<11.0.0)",
  "numpy (>=1.26.0,<2.0.0)",
  "paddlepaddle (>=3.3.0,<4.0.0)",
]

[dependency-groups]
dev = ["pylint (>=4.0.0,<5.0.0)", "mypy (>=1.18.2,<2.0.0)", "basedpyright (<3.12)", "pyinstaller (>=6.16.0,<7.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.basedpyright]
reportUnknownMemberType=false
reportUnknownLambdaType=false
reportUnusedFunction=true
reportMissingTypeStubs=false
reportExplicitAny=false
reportAny=false
exclude = ["tests/"]
ignore = ["tests/"]

[[tool.mypy.overrides]]
module = ["msgpack.*"]
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["paddleocr.*", "paddle.*", "pytesseract.*", "numpy.*", "PIL.*"]
ignore_missing_imports = true

[tool.poetry.scripts]
build-executable = "build:main"
